+++ examples/find_tags.py
@@ -1,6 +1,6 @@
-'''
+"""
 Find RuuviTags
-'''
+"""
 
 from ruuvitag_sensor.ruuvi import RuuviTagSensor
 import ruuvitag_sensor.log