+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### [Unreleased]
 * Fix hcitool subprocess closing
+* Fix data format selection
+* Move broadcasted raw data handling to an own file
+* Remove obsolete handling for r/ in Data Format 2/4 validate
 
 ## [0.12.0] - 2019-02-15
 * Changed RuuviTagReactive's time-value from local to UTC